+++ VulkanGraphics/Object.cpp
@@ -4,7 +4,7 @@ Object::Object()
 {
 }
 
-Object::Object(VulkanEngine& engine, std::shared_ptr<TextureManager> textureManager, std::vector<Vertex> vertexdata, std::vector<uint16_t> indicesdata, MeshTextures textures, int cubemap, VkDescriptorSetLayout& descriptorSetLayout, int renderBit)
+Object::Object(VulkanEngine& engine, std::shared_ptr<TextureManager> textureManager, std::vector<Vertex> vertexdata, std::vector<uint32_t> indicesdata, MeshTextures textures, int cubemap, VkDescriptorSetLayout& descriptorSetLayout, int renderBit)
 {
 	//ObjectMesh = std::make_shared<Mesh>(Mesh(renderer, textureManager, vertexdata, indicesdata, textures, cubemap, descriptorSetLayout, renderBit));
 //	renderer.AddDrawableMesh(ObjectMesh);
